+++ resolwe/flow/tests/test_manager.py
@@ -43,7 +43,8 @@ class TestManager(ProcessTestCase):
         )
         spawned_process = Process.objects.filter(slug='test-save-file').latest()
         # Patch the process to create Entity, so its bahaviour can be tested.
-        spawned_process.flow_collection = 'test-schema'
+        spawned_process.entity_type = 'test-schema'
+        spawned_process.entity_descriptor_schema = 'test-schema'
         spawned_process.save()
 
         process = Process.objects.filter(slug='test-spawn-new').latest()